Sync: ignore lastModifiedDate when calculating event hash.
I can't manage to figure out why, but it looks like that with events
that don't change (e.g. a CHANGE/ADD in the journal to an existing
identical event) the lastModifiedDate is not correct when fetching the
event immediately after the update.

I've also seen similar behaviour happen with other events too. I'm not
sure if it's the same issue, but it may not be limited to the events
described above, and I just notice it more with them because they are
significantly more common.