SyncManager: don't do a full sync in the background.
We were taking too long and causing iOS to cancel our background
fetching. We can't enable push again until we are significantly faster,
or potentially just take another route.

This is kind of a revert of fcc7ec1ea261700ce31e3cdd7809c15b15982d12. We
changed quite a few things afterwards so it can't be a real revert, but
it's essentially rolling back that commit.