Store: change encryption key's name in the store
I haven't managed to figure out exactly why, but it looks like the name
"key" is causing issues. After changing this to "encryptionKey" the
issue with the encryptionKey not always being persisted is fixed.

What was actually happening is that redux-persist was loading the
correct encryption key but then immediately afterwards was persisting
the default value ( { key: null } ) thus losing the encryption key.

As said, I wasn't able to figure out what's going on exactly, but it
seems like upgrading redux-persist + changing this fixed things.